# Startup `flutter` faster (cache git calls)