--- a/packages/flutter_tools/lib/src/commands/channel.dart
+++ b/packages/flutter_tools/lib/src/commands/channel.dart
@@ -182,7 +182,7 @@ class ChannelCommand extends FlutterCommand {
     } else {
       // Remove the version check stamp, since it could contain out-of-date
       // information that pertains to the previous channel.
-      await FlutterVersion.resetFlutterVersionFreshnessCheck();
+      await FlutterVersion.resetFlutterVersionCache();
     }
   }
 }