This PR makes `flutter` (as measured with `flutter test` directly in the `flutter` directory which has no `test` directory, thus just writing "there's no `test` directory) startup faster by caching `git` calls.

When starting `flutter` a number of `git` calls are made, which is quite expensive. This PR introduces caching of these calls.
Caching is deleted when also creating a new snapshot (i.e. it's guarded by the same mechanism, which is the git hash).

This does mean that these things can change:
* Two branches on the same hash will have an undetected stale cache (in that it will think it's on one of the branch when it could be on another). I'll say this is very unlikely from having any real effect. Logically we might report crashes or whatever as if on the "main" branch when we're really on our own not-yet-committed-to copy of main, but logically this already happens when we haven't created a branch yet but have made changes to the code.
* If official tags are created for old versions where the user cached the old answer (e.g. no tags, user uses `flutter` and caches that, a tag is then created, the user fetches (but stays on the old hash) --- git would now know about the new tag, but caching isn't updated. Again I'll say this doesn't matter and any crash reports or whatever would have to handle the case anyway because not everyone has necessarily fetched.

To me, neither of these seems like big deals though.
The startup changes as measured on my machines is this:

**Linux**
| Before on master | Before not on master | With PR |
| --- | --- | --- |
| 0.614 | 0.606 | 0.484 |
| 0.655 | 0.602 | 0.482 |
| 0.69 | 0.605 | 0.489 |
| 0.669 | 0.595 | 0.474 |
| 0.647 | 0.598 | 0.493 |
| 0.672 | 0.584 | 0.483 |
| 0.638 | 0.602 | 0.483 |
| 0.67 | 0.614 | 0.489 |
| 0.67 | 0.593 | 0.48 |
| 0.677 | 0.609 | 0.489 |

While not important for this PR I'll note that it's faster to startup `flutter` when not on a branch called master:
```
Difference at 95.0% confidence
        -0.0594 +/- 0.0157866
        -8.99727% +/- 2.39118%
        (Student's t, pooled s = 0.0168015)
```

This PR improves things (compared to now on branch called master):

```
Difference at 95.0% confidence
        -0.1756 +/- 0.0151538
        -26.598% +/- 2.29533%
        (Student's t, pooled s = 0.016128)
```

This PR improves things (compared to now on branch not called master):

```
Difference at 95.0% confidence
        -0.1162 +/- 0.00681882
        -19.3409% +/- 1.13496%
        (Student's t, pooled s = 0.00725718)
```

I.e. it's somewhere from 20 to 25% faster, cutting off something like 110 to 175 ms.

**Windows** (Google issued with whatever security software)
| Before on master | Before not on master | Now |
| --- | --- | --- |
| 3.2498843 | 3.2052244 | 2.323404 |
| 3.2471592 | 3.2368394 | 2.3255623 |
| 3.2630213 | 3.2279151 | 2.3146872 |
| 3.2694762 | 3.2593181 | 2.2819018 |
| 3.2471021 | 3.2587005 | 2.3097854 |
| 3.2608728 | 3.2641268 | 2.3177364 |
| 3.2796176 | 3.2531503 | 2.2960537 |
| 3.2498695 | 3.2752845 | 2.3057254 |
| 3.2417869 | 3.2853195 | 2.3059499 |
| 3.2782214 | 3.2681917 | 2.3037698 |

This PR improves things:

```
Difference at 95.0% confidence
        -0.944949 +/- 0.0180964
        -29.0449% +/- 0.556228%
        (Student's t, pooled s = 0.0192597)
```

I.e. it's almost 30% faster, cutting off almost a full second.

A more detailed analysis is available internally at go/FlutterStartupTimeAnalysis (which also discussed future PRs).