@jonahwilliams 2022-09-20
For the actual caching behavior you've outlined, I would defer to our resident git authority @christopherfujino .

I would say that in general I'm okay with trying this out because we have a lot of runway before our next release, but in general it does strike me as the change most likely to introduce a bug.