@jensjoha 2022-09-23
> However, if we ARE on master, I don't think we want to cache this call, as there could have been a newer tag added after we checked to a commit that is an ancestor of our current HEAD, and we would want the Flutter version to report the commit as being more recent than that.

I don't understand this part. Either I don't understand what you're saying, or I don't understand why it should be the case (the way I'm reading it we should suddenly report an old version as being newer?).

Also it's not clear to me which command you think shouldn't be cached.
The cached ones are these:

- `git -c log.showSignature=false log -n 1 --pretty=format:%H`. This command does not seem to fetch anything from the network (having `[url "git@github.com:"]\n    insteadof = https://github.com/` in my `.gitconfig` asks for my password for stuff like fetching, and this command does not ask for my password so I assume it doesn't access the network at all).
- `git rev-parse --abbrev-ref --symbolic @{upstream}`. This command does not seem to fetch anything from the network.
- `git ls-remote --get-url origin`. This command does not seem to fetch anything from the network.
- `git -c log.showSignature=false log HEAD -n 1 --pretty=format:%ad --date=iso`. This command does not seem to fetch anything from the network.
- `git rev-parse --abbrev-ref HEAD`. This command does not seem to fetch anything from the network.
- `git tag --points-at hash` (e.g. `git tag --points-at cbfb8fa2b2fe625db03535b313cd3f5957680ac1`). This command does not seem to fetch anything from the network.
- `git describe --match *.*.* --long --tags hash` (e.g. `git describe --match *.*.* --long --tags cbfb8fa2b2fe625db03535b313cd3f5957680ac1`). This command does not seem to fetch anything from the network.

So none of the cached commands access the network. So the only way the answers can become "externally out-of-date" is by other commands accessing the network:

 - `git -c log.showSignature=false log -n 1 --pretty=format:%ar` is not cached because it gives a relative time (e.g. `2 hours ago`) but does not access the network.
 - `git fetch https://github.com/flutter/flutter.git --tags -f`. This command accesses the internet --- specifically to fetch tags. We should probably delete the cache after this call, so I've added that.
- The user could manually run `git fetch` or similar. I don't think we can really guard against that. We could add a time-guard like in `_getLatestAvailableFlutterDate` (or possibly just delete the cache when actually getting the latest flutter date anyway).