@christopherfujino 2022-09-29
> > However, if we ARE on master, I don't think we want to cache this call, as there could have been a newer tag added after we checked to a commit that is an ancestor of our current HEAD, and we would want the Flutter version to report the commit as being more recent than that.
> 
> I don't understand this part. Either I don't understand what you're saying, or I don't understand why it should be the case (the way I'm reading it we should suddenly report an old version as being newer?).

Consider that you check out the master branch at commit abcd, which is 10 commits past the last tagged release, version `1.0.0`. The output of `git describe --match *.*.* --long --tags abcd` would be something like `1.0.0~10`.

Now consider we tag a commit 5 commits past version 1 and 5 commits before commit abcd as version 2.0.0. *Now*, the result of `git describe --match *.*.* --long --tags abcd` would be `2.0.0~5`.