@jensjoha 2022-09-30
> > > However, if we ARE on master, I don't think we want to cache this call, as there could have been a newer tag added after we checked to a commit that is an ancestor of our current HEAD, and we would want the Flutter version to report the commit as being more recent than that.
> > 
> > 
> > I don't understand this part. Either I don't understand what you're saying, or I don't understand why it should be the case (the way I'm reading it we should suddenly report an old version as being newer?).
> 
> Consider that you check out the master branch at commit abcd, which is 10 commits past the last tagged release, version `1.0.0`. The output of `git describe --match *.*.* --long --tags abcd` would be something like `1.0.0~10`.
> 
> Now consider we tag a commit 5 commits past version 1 and 5 commits before commit abcd as version 2.0.0. _Now_, the result of `git describe --match *.*.* --long --tags abcd` would be `2.0.0~5`.

This would require the user to fetch new tags though. Meaning there will always be a window in which they see it as `1.0.0~10` even after it has been tagged, and thus the scenario already happens. I've inserted an invalidation in `fetchRemoteFrameworkCommitDate()` now (which is what's called from `_getLatestAvailableFlutterDate` if it's been too long since the last check) meaning that such cases should always be noticed within ~3 days.
In - I'm guessing - all likely scenarios (i.e. when the user doesn't run `git fetch` manually) this makes the tag description give the "wrong" result at exactly the same times as before.